pyproject.toml
[tool.poetry]
name = "jwt-tester"
version = "0.1.0"
description = ""
authors = ["suk-6 <me@suk.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
black = "^24.10.0"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
